gogo

+++ printf/ft_get_md.c
@@ -6,14 +6,14 @@
 /*   By: ukwon <ukwon@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/28 19:41:07 by ukwon             #+#    #+#             */
-/*   Updated: 2020/10/30 14:53:20 by ukwon            ###   ########.fr       */
+/*   Updated: 2020/11/01 18:59:22 by ukwon            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 #include "libft/libft.h"
 
-void			get_md(t_flag *flag, int p)
+void			get_md(t_flag *flag, long long int p)
 {
 	int i;
 	int	p_i;
