+++ printf/libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: ukwon <ukwon@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/08 03:40:50 by ukwon             #+#    #+#             */
-/*   Updated: 2020/10/30 14:56:23 by ukwon            ###   ########.fr       */
+/*   Updated: 2020/11/01 17:43:52 by ukwon            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -41,7 +41,7 @@ char			*ft_strchr(const char *s, int c);
 char			*ft_strrchr(const char *s, int c);
 int				ft_memcmp(const void *s1, const void *s2, size_t n);
 void			*ft_memchr(const void *s, int c, size_t n);
-char			*ft_itoa(int n);
+char			*ft_itoa(long long int n);
 char			*ft_strmapi(char *s, char (*f)(unsigned int, char));
 void			ft_putchar_fd(char c, int fd);
 void			ft_putstr_fd(char const *s, int fd);
